requirements.txt
openai>=1.2.0
streamlit
python-dotenv
pytest